Extract JavaType class and some utils.
