+++ www/js/project_view.js
@@ -8,7 +8,8 @@
  */
 
 import ASTDiff from './ast-diff.js';
-import Cell from './cell.js';
+import JavaType from './java-type.js';
+import { first, last } from './utils.js';
 
 
 var VALID_STEP_SIZES = d3.set(['hour', 'day', 'month', 'week']);
@@ -63,134 +64,6 @@ function createTable2(filter) {
 
 
 /*=== Classes ===*/
-
-
-
-
-/**
- * Class: JavaType
- *
- * Represents a type in Java.
- * Note: There probably should be a JavaEntity,
- * with methods having a JavaType parent, and JavaType arguments;
- * and for there to be a DataEntry class that contains cells. But that sounds
- * like too much effort.
- *
- * Create an array of these and you get a y-axis!
- */
-function JavaType(name) {
-  /* Instantiate a new Cell object if called without `new`. */
-  if (!(this instanceof JavaType)) {
-    return new JavaType(name);
-  }
-
-  /* TODO: primitive types? */
-  /* TODO: generics? */
-  var sides = name.split('#');
-  var generics_components = sides[0].split('<');
-  var components = generics_components[0].split('.');
-
-  /* Max adds per type. */
-  this._largest = null;
-
-  this.className = components.pop();
-  this.package = components.join('.');
-  this.methodName = sides[1] || '';
-
-  if (generics_components.length > 1) {
-    this.className += "<" + generics_components[1];
-  }
-
-  this.cells = [];
-};
-
-Object.defineProperties(JavaType.prototype, {
-  /**
-   * Returns the fully-qualified name (package + class name) of this type.
-   */
-  name: {
-    get: function () {
-      if (this.package) {
-        return this.package + '.' + this.shortName;
-      }
-
-      return this.shortName;
-    }
-  },
-
-  /**
-   * The short name is simply the class name, plus its arguments, if they
-   * exist.
-   */
-  shortName: {
-    get: function () {
-      if (this.methodName) {
-        return this.className + '#' + this.methodName;
-      }
-      return this.className;
-    }
-  },
-
-  /**
-   * Alias for name.
-   */
-  fullyQualifiedName: {
-    get: function () {
-      return this.name;
-    }
-  },
-
-  /**
-   * Adds a data cell.
-   */
-  addCell: {
-    value: function (cell) {
-      assert(cell instanceof Cell);
-      return this.cells.push(cell);
-    }
-  },
-
-  /**
-   * Adds an AST diff to the current type.
-   */
-  addDiff: {
-    value: function (diff, startDate, endDate) {
-      if (!this.cells.length || !last(this.cells).isAcceptableDiff(diff)) {
-        /* Automatically add a new cell, if needed. */
-        this.addCell(new Cell(startDate, endDate, this.name));
-      }
-      last(this.cells).addDiff(diff);
-    }
-  },
-
-  /**
-   * Returns the number of commits in the largest cell.
-   */
-  numberOfObservationsInLargestCell: {
-    get: function () {
-      if (this._largest) {
-        return this._largest;
-      }
-
-      /* Calculate it! */
-      var maxCell = d3.max(this.cells, function (cell) {
-        return cell.numberOfObservations;
-      });
-
-      this._largest = maxCell;
-      return this._largest;
-    }
-  },
-
-  /**
-   * toString() will simply return the fully-qualified name.
-   */
-  toString: {
-    value: function () {
-      return this.fullyQualifiedName;
-    }
-  }
-});
 
 
 /**
@@ -1019,21 +892,7 @@ function cellWidthFromScale(cell, scale) {
   return Math.ceil(bigger - smaller) + 1;
 }
 
-function first(array) {
-  assert(array instanceof Array);
-  if (array.length < 1) {
-    throw new Error('Cannot get the first item of an empty array; is the project empty?');
-  }
-  return array[0];
-}
 
-function last(array) {
-  assert(array instanceof Array);
-  if (array.length < 1) {
-    throw new Error('Cannot get the first item of an empty array; is the project empty?');
-  }
-  return array[array.length - 1];
-}
 
 /*=== Predicates used in assertions and checks ===*/
